// Sebo_UX/src/app/interfaces/user-profile.ts
export interface UserProfile {
    _creationTime: Date;
    _id: String;
    about: String;
    country: String;
    created_by: String;
    date_of_birth: String;
    email: String;
    family_name: String;
    given_name: String;
    hobbies: String;
    interests: String;
    name: String;
    phone_number: String;
    profile_thumbnail: String;
    social_image_url: String;
    username: String;

}